pyproject.toml
[build-system]
requires = ["hatchling>=1.26.3", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "chessimg2pos"
description = "Recognize chess positions from board images using deep learning"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
keywords = ["chess", "FEN", "deep learning", "PyTorch", "computer vision"]
authors = [
  { name = "Marco Di Ciò", email = "marcodicio.md@gmail.com" }
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Intended Audience :: Education",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Topic :: Games/Entertainment :: Board Games",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Software Development :: Libraries :: Python Modules"
]
dependencies = [
  "torch>=2.0.0",
  "torchvision>=0.15.0",
  "numpy>=1.24.0",
  "matplotlib>=3.7.0",
  "Pillow>=10.0.0",
  "pandas>=2.0.0",
  "requests>=2.28.0"
]
dynamic = ["version"]

[project.urls]
Homepage = "https://github.com/mdicio/chessimg2pos"
Repository = "https://github.com/mdicio/chessimg2pos"
Issues = "https://github.com/mdicio/chessimg2pos/issues"

[tool.hatch.version]
path = "src/chessimg2pos/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/chessimg2pos"]

